+++ azure-webapp-maven-plugin/src/main/java/com/microsoft/azure/maven/webapp/parser/V2ConfigurationParser.java
@@ -14,25 +14,26 @@ import com.microsoft.azure.maven.webapp.AbstractWebAppMojo;
 import com.microsoft.azure.maven.webapp.configuration.Deployment;
 import com.microsoft.azure.maven.webapp.configuration.OperatingSystemEnum;
 import com.microsoft.azure.maven.webapp.configuration.RuntimeSetting;
+import com.microsoft.azure.maven.webapp.validator.AbstractConfigurationValidator;
 import org.apache.maven.model.Resource;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.codehaus.plexus.util.StringUtils;
+
 import java.util.List;
 import java.util.Locale;
 
 public class V2ConfigurationParser extends ConfigurationParser {
-    public V2ConfigurationParser(AbstractWebAppMojo mojo) {
-        super(mojo);
+
+    public V2ConfigurationParser(AbstractWebAppMojo mojo, AbstractConfigurationValidator validator) {
+        super(mojo, validator);
     }
 
     @Override
     protected OperatingSystemEnum getOs() throws MojoExecutionException {
+        validate(validator.validateOs());
         final RuntimeSetting runtime = mojo.getRuntime();
         final String os = runtime.getOs();
         if (runtime.isEmpty()) {
             return null;
-        } else if (StringUtils.isEmpty(os)) {
-            throw new MojoExecutionException("Pleas configure the <os> of <runtime> in pom.xml.");
         }
         switch (os.toLowerCase(Locale.ENGLISH)) {
             case "windows":
@@ -42,22 +43,20 @@ public class V2ConfigurationParser extends ConfigurationParser {
             case "docker":
                 return OperatingSystemEnum.Docker;
             default:
-                throw new MojoExecutionException("The value of <os> is not correct, supported values are: windows, " +
-                    "linux and docker.");
+                return null;
         }
     }
 
     @Override
     protected Region getRegion() throws MojoExecutionException {
+        validate(validator.validateRegion());
         final String region = mojo.getRegion();
-        if (!StringUtils.isEmpty(region) && Region.findByLabelOrName(region) == null) {
-            throw new MojoExecutionException("The value of <region> is not supported, please correct it in pom.xml.");
-        }
         return Region.fromName(region);
     }
 
     @Override
     protected RuntimeStack getRuntimeStack() throws MojoExecutionException {
+        validate(validator.validateRuntimeStack());
         final RuntimeSetting runtime = mojo.getRuntime();
         if (runtime == null || runtime.isEmpty()) {
             return null;
@@ -67,13 +66,8 @@ public class V2ConfigurationParser extends ConfigurationParser {
 
     @Override
     protected String getImage() throws MojoExecutionException {
+        validate(validator.validateImage());
         final RuntimeSetting runtime = mojo.getRuntime();
-        if (runtime == null) {
-            throw new MojoExecutionException("Please configure the <runtime> in pom.xml.");
-        }
-        if (StringUtils.isEmpty(runtime.getImage())) {
-            throw new MojoExecutionException("Please config the <image> of <runtime> in pom.xml.");
-        }
         return runtime.getImage();
     }
 
@@ -102,22 +96,15 @@ public class V2ConfigurationParser extends ConfigurationParser {
 
     @Override
     protected WebContainer getWebContainer() throws MojoExecutionException {
+        validate(validator.validateWebContainer());
         final RuntimeSetting runtime = mojo.getRuntime();
-        if (runtime == null) {
-            throw new MojoExecutionException("Pleas config the <runtime> in pom.xml.");
-        }
         return runtime.getWebContainer();
     }
 
     @Override
     protected JavaVersion getJavaVersion() throws MojoExecutionException {
+        validate(validator.validateJavaVersion());
         final RuntimeSetting runtime = mojo.getRuntime();
-        if (runtime == null) {
-            throw new MojoExecutionException("Pleas config the <runtime> in pom.xml.");
-        }
-        if (runtime.getJavaVersion() == null) {
-            throw new MojoExecutionException("The configuration <javaVersion> in pom.xml is not correct.");
-        }
         return runtime.getJavaVersion();
     }