+++ azure-webapp-maven-plugin/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 All notable changes to the "Maven Plugin for Azure App Service" will be documented in this file.
 - [Change Log](#change-log)
+  - [1.7.0](#170)
   - [1.6.0](#160)
   - [1.5.4](#154)
   - [1.5.3](#153)
@@ -13,6 +14,14 @@ All notable changes to the "Maven Plugin for Azure App Service" will be document
   - [1.2.0](#120)
   - [1.1.0](#110)
   - [1.0.0](#100)
+
+## 1.7.0
+- Support customize java options for Windows Java SE app service.[#640](https://github.com/microsoft/azure-maven-plugins/issues/640)
+- Support update app service plan for existing app service.[PR#677](https://github.com/microsoft/azure-maven-plugins/pull/677)
+- Generate version tag if `<plugin>` is absent while `webapp:config`.[#660](https://github.com/microsoft/azure-maven-plugins/issues/660)
+- Optimization prompt messages.[#667](https://github.com/microsoft/azure-maven-plugins/issues/667)
+- Fix: Plugin can't get correct subscription in azure cloud shell.[#628](https://github.com/microsoft/azure-maven-plugins/issues/628)
+- Fix: Plugin can't get client id when user login azure cli with service principal.[#125](https://github.com/microsoft/azure-maven-plugins/issues/125)
 
 ## 1.6.0
 - Support Java 11 AppService.[PR#606](https://github.com/Microsoft/azure-maven-plugins/pull/606)